% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_move.r
\name{cat_move}
\alias{cat_move}
\alias{cat_game}
\title{View game in command-line terminal}
\usage{
cat_move(game, move = NULL, ...)

cat_game(game, ..., fps = 1)
}
\arguments{
\item{game}{A list containing a parsed ppn game (as parsed by \code{\link[=read_ppn]{read_ppn()}})}

\item{move}{Which move to cat game state (after the move, will use \code{game$dfs[[move]]})
unless \code{NULL} in which case will print the game state after the last move.}

\item{...}{Passed to \code{\link[ppcli:cat_piece]{ppcli::cat_piece()}}.}

\item{fps}{Frames per second.}
}
\description{
\code{cat_move()} prints a plaintext diagram of a single move to the terminal.
\code{cat_game()} prints a plaintext "animation" of every move to the terminal.
}
