% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_parser.r
\name{default_parser}
\alias{default_parser}
\alias{alquerque_parser}
\alias{go_parser}
\alias{marble_parser}
\alias{morris_parser}
\title{Movetext parsers}
\usage{
default_parser(
  movetext = character(),
  metadata = list(),
  ...,
  scale_factor = NULL,
  default_system = "piecepack"
)

alquerque_parser(movetext = character(), metadata = list(), ...)

go_parser(movetext = character(), metadata = list(), ...)

marble_parser(movetext = character(), metadata = list(), ...)

morris_parser(movetext = character(), metadata = list(), ...)
}
\arguments{
\item{movetext}{A named character vector of move text.}

\item{metadata}{A named list of metadata}

\item{...}{Should be empty for \code{default_parser()}.  Otherwise passed to \code{default_parser()}.}

\item{scale_factor}{How many inches a coordinate unit is worth.}

\item{default_system}{Default game system to use for setup functions.}
}
\value{
A named list of data frames with parsed game states.
}
\description{
\code{default_parser()} is the default PPN movetext parser.
}
\examples{
l <- read_ppn(system.file("ppn/relativity.ppn", package = "ppn"),
              parse = FALSE)[[1]]
game <- default_parser(l$movetext, l$metadata)
}
